update script to check depends and create folders

+++ testplatforms.sh
@@ -14,13 +14,16 @@ for rver in "${version[@]}"
 do
     LIBLOC="$HOME/R/r-${rver}"
     echo "Setting LIBLOC = $LIBLOC"
+    if [ ! -d $LIBLOC ]; then
+        mkdir -p $LIBLOC
+    fi
 
     R_LIBS_USER="$HOME/R/r-${rver}"
     echo "Setting R_LIBS_USER = \"$HOME/R/r-${rver}\""
 
     rversion="$HOME/src/svn/r-${rver}/R/bin/R --vanilla"
     
-#    ${rversion} -e "install.packages(c('knitr', 'testthat', 'remotes', 'stringr'))"
+    ${rversion} -e "deps <- c('knitr', 'testthat', 'remotes', 'stringr'); if (!all(deps %in% rownames(installed.packages()))) install.packages(deps)"
 
     cd $LIBLOC
 
@@ -32,6 +35,12 @@ do
     echo "** _R_CHECK_FORCE_SUGGESTS_=FALSE ${rversion} CMD check ${TARBALL}"
     _R_CHECK_FORCE_SUGGESTS_=FALSE ${rversion} CMD check ${TARBALL}
 
+    if [ $? -ne 0 ]; then
+        echo "Unable to check package without errors"
+        exit 2
+    fi
+
     echo "** Removing tarball... rm ${TARBALL}"
     rm ${TARBALL}
+
 done
